task 3

+++ 0x0A-argc_argv/3-mul.c
@@ -14,7 +14,7 @@ int main(int argc, char *argv[])
 	{
 		int i;
 
-		i = *argv[1] * *argv[2];
+		i = (int)*argv[1] * (int)*argv[2];
 		printf("%d\n", i);
 
 		return (0);
